test: relocate PR-review regressions into domain files (#684 item 4)
Move 32 tests out of history-named test_pr540_review.py and
test_pr635_review.py into their domain homes:

- test_scanner.py    scan-scope / global-search / prerequisites (6)
- test_graph.py      custom-graph + show_tree, marked integration (4)
- test_push.py       push-injection guards (3)
- test_formatting.py force-color, marked integration (1)
- test_cli_core.py   TUI + parser-flag + resolve-path checks (13)

Both history-named shells deleted via git rm. Test bodies, names,
and assertions preserved verbatim; helpers (_make_skill, _make_registry)
copied where needed. Fast tier: 796 passed. No failures introduced.
