--- tests/test_pr635_review.py
@@ -1,639 +1,0 @@
-"""PR #635 Sandbox Review — RED / GREEN tests for gaia scan, push, tree --custom, graph --custom.
-
-RED tests   : verify the new behaviour was NOT present on origin/main
-GREEN tests : verify the new behaviour IS present on the PR branch
-
-Each test is self-contained and uses tmp_path / monkeypatch to avoid side-effects.
-"""
-
-import json
-import os
-from types import SimpleNamespace
-
-import pytest
-
-from gaia_cli.scanner import scan_skill_mds, _skill_search_dirs
-from gaia_cli.treeManager import show_tree
-from gaia_cli import graph as graph_mod
-pytestmark = [pytest.mark.integration]
-
-
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-def _write(path, text):
-    os.makedirs(os.path.dirname(path), exist_ok=True)
-    with open(path, "w", encoding="utf-8") as f:
-        f.write(text)
-
-
-def _make_skill(tmp_path, rel_dir, skill_id, *, name=None, description="", prerequisites=None):
-    """Create a minimal skill dir with a SKILL.md under `tmp_path / rel_dir / skill_id`."""
-    skill_dir = tmp_path / rel_dir / skill_id
-    skill_dir.mkdir(parents=True, exist_ok=True)
-    fm_lines = ["---"]
-    fm_lines.append(f"name: {name or skill_id}")
-    if description:
-        fm_lines.append(f"description: {description}")
-    if prerequisites:
-        fm_lines.append(f"prerequisites: {json.dumps(prerequisites)}")
-    fm_lines.append("---")
-    fm_lines.append(f"# {name or skill_id}")
-    (skill_dir / "skill.md").write_text("\n".join(fm_lines), encoding="utf-8")
-    return skill_dir
-
-
-def _make_registry(root, *, skills=None):
-    """Create a minimal registry dir structure for graph tests."""
-    registry = root / "registry"
-    registry.mkdir(parents=True, exist_ok=True)
-    graph_data = {
-        "version": "test",
-        "generatedAt": "2026-06-08",
-        "skills": skills or [],
-    }
-    (registry / "gaia.json").write_text(json.dumps(graph_data), encoding="utf-8")
-    (registry / "named-skills.json").write_text(
-        json.dumps({"buckets": {}}), encoding="utf-8"
-    )
-    return root
-
-
-# ═══════════════════════════════════════════════════════════════════════════
-# RED TESTS — scenarios that FAIL on origin/main, PASS on PR branch
-# ═══════════════════════════════════════════════════════════════════════════
-
-
-class TestRed_ScanScopeToProjectRoot:
-    """RED #1: scan_skill_mds scopes to project root by default.
-
-    On main: scan_skill_mds(root) walks root/.. (the parent), potentially
-    leaking across sibling project dirs.
-    On branch: walks only root itself when global_search=False (default).
-    """
-
-    def test_default_scan_does_not_walk_parent(self, tmp_path):
-        """A skill in a sibling directory should NOT be discovered by default scan."""
-        # Create a skill in a sibling directory (outside the project root)
-        sibling = tmp_path / "sibling-project"
-        _make_skill(tmp_path, os.path.join("sibling-project", ".agents", "skills"),
-                     "leaked-skill", name="Leaked Skill")
-
-        # Create our project root (no skills)
-        project_root = tmp_path / "my-project"
-        project_root.mkdir(parents=True)
-
-        # On main: scan_skill_mds(root=my-project) walked root/.. which IS tmp_path,
-        # so it would find sibling-project/.agents/skills/leaked-skill.
-        # On branch: scan_skill_mds(root=my-project) walks only my-project.
-        results = scan_skill_mds(root=str(project_root), global_search=False)
-        found_ids = {r["id"] for r in results}
-
-        # This MUST NOT find the sibling's skill (the RED condition on main).
-        assert "leaked-skill" not in found_ids, (
-            "Default scan should not walk the parent directory and find sibling skills"
-        )
-
-
-class TestRed_GlobalSearchDirsExcluded:
-    """RED #2: _skill_search_dirs excludes global dirs when global_search=False.
-
-    On main: ~/.agents/skills was always included.
-    On branch: only included when global_search=True.
-    """
-
-    def test_global_dirs_excluded_by_default(self, tmp_path, monkeypatch):
-        """When global_search=False, home-based dirs should NOT appear."""
-        home = tmp_path / "fake-home"
-        global_skills = home / ".agents" / "skills"
-        global_skills.mkdir(parents=True)
-
-        monkeypatch.setenv("HOME", str(home))
-
-        dirs = _skill_search_dirs(root=str(tmp_path), global_search=False)
-        real_dirs = [os.path.realpath(d) for d in dirs]
-        assert os.path.realpath(str(global_skills)) not in real_dirs, (
-            "Global dirs should be excluded when global_search=False"
-        )
-
-    def test_global_dirs_included_when_requested(self, tmp_path, monkeypatch):
-        """When global_search=True, home-based dirs should appear (if they exist)."""
-        home = tmp_path / "fake-home"
-        global_skills = home / ".agents" / "skills"
-        global_skills.mkdir(parents=True)
-
-        monkeypatch.setenv("HOME", str(home))
-
-        dirs = _skill_search_dirs(root=str(tmp_path), global_search=True)
-        real_dirs = [os.path.realpath(d) for d in dirs]
-        assert os.path.realpath(str(global_skills)) in real_dirs, (
-            "Global dirs should be included when global_search=True"
-        )
-
-
-class TestRed_PrerequisitesField:
-    """RED #3: scan_skill_mds result includes 'prerequisites' key.
-
-    On main: result dicts had no 'prerequisites' key.
-    On branch: 'prerequisites' key is present (defaults to [] when absent).
-    """
-
-    def test_result_has_prerequisites_key(self, tmp_path):
-        """Every scanned skill result must have a 'prerequisites' key."""
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "skill-a",
-                     name="Skill A", prerequisites=["skill-b"])
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "skill-b",
-                     name="Skill B")
-
-        results = scan_skill_mds(root=str(tmp_path))
-        assert len(results) >= 2
-
-        for result in results:
-            assert "prerequisites" in result, (
-                f"Skill {result['id']} is missing the 'prerequisites' key"
-            )
-
-    def test_prerequisites_default_to_empty_list(self, tmp_path):
-        """Skills without frontmatter prerequisites get an empty list."""
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "no-prereqs",
-                     name="No Prereqs")
-
-        results = scan_skill_mds(root=str(tmp_path))
-        assert len(results) == 1
-        assert results[0]["prerequisites"] == [], (
-            "Missing frontmatter prerequisites should default to []"
-        )
-
-
-class TestRed_PushInjectsUnmatchedSkills:
-    """RED #4: push_command injects unmatched local custom skills into batch.
-
-    On main: skills not already in batch_proposed_ids or batch_known_ids were DROPPED.
-    On branch: they are injected into batch['proposedSkills'].
-
-    This test verifies the injection logic directly rather than running the
-    full push_command (which requires interactive input and git remotes).
-    """
-
-    def test_injection_logic_adds_missing_skills_to_batch(self, tmp_path):
-        """Simulate the push_command injection loop: skills not in batch should be added."""
-        # Simulate what push_command does
-        batch = {
-            "sourceRepo": "test/repo",
-            "proposedSkills": [
-                {"id": "already-proposed", "name": "Already", "type": "basic",
-                 "description": "Already in batch", "sourceRepo": "test/repo",
-                 "lifecycle": "pending"}
-            ],
-            "knownSkills": [
-                {"skillId": "already-known"}
-            ],
-        }
-        batch_proposed_ids = {s["id"] for s in batch.get("proposedSkills", [])}
-        batch_known_ids = {s["skillId"] for s in batch.get("knownSkills", [])}
-
-        # Simulate installed_skills (what scan_skill_mds returns)
-        installed_skills = [
-            {"id": "already-proposed", "name": "Already Proposed", "description": ""},
-            {"id": "already-known", "name": "Already Known", "description": ""},
-            {"id": "brand-new-skill", "name": "Brand New", "description": "A new skill"},
-        ]
-
-        # This is the injection logic from the PR branch
-        for sk in installed_skills:
-            cid = sk["id"]
-            if cid not in batch_proposed_ids and cid not in batch_known_ids:
-                batch.setdefault("proposedSkills", []).append({
-                    "id": cid,
-                    "name": sk.get("name", cid),
-                    "type": "basic",
-                    "description": sk.get("description", f"Local custom skill {cid}"),
-                    "sourceRepo": batch.get("sourceRepo", "unknown"),
-                    "lifecycle": "pending",
-                })
-                batch_proposed_ids.add(cid)
-
-        # Verify: brand-new-skill should have been injected
-        proposed_ids = {s["id"] for s in batch["proposedSkills"]}
-        assert "brand-new-skill" in proposed_ids, (
-            "Unmatched local custom skill should be injected into proposedSkills"
-        )
-        # Verify no duplicates
-        assert len([s for s in batch["proposedSkills"] if s["id"] == "already-proposed"]) == 1, (
-            "Already-proposed skill should not be duplicated"
-        )
-
-
-class TestRed_WriteGraphArtifactCustom:
-    """RED #5: write_graph_artifact(custom=True) uses local scan, not load_graph().
-
-    On main: no 'custom' parameter — load_graph() always used.
-    On branch: custom=True bypasses load_graph().
-    """
-
-    def test_custom_graph_uses_local_scan(self, tmp_path, monkeypatch):
-        """custom=True should build graph from scan_skill_mds, not load_graph()."""
-        root = _make_registry(tmp_path, skills=[
-            {"id": "registry-skill", "name": "Registry Skill", "type": "basic",
-             "level": "1★", "prerequisites": []},
-        ])
-
-        # Create a local custom skill that is NOT in the registry
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "local-only",
-                     name="Local Only", description="A local-only custom skill")
-        monkeypatch.chdir(tmp_path)
-
-        out_path, _ = graph_mod.write_graph_artifact(
-            root, fmt="json", custom=True
-        )
-
-        data = json.loads(out_path.read_text(encoding="utf-8"))
-
-        # The custom graph should contain local-only skill from scan
-        node_ids = {n["id"] for n in data["nodes"]}
-        assert "local-only" in node_ids, (
-            "custom=True graph should include locally scanned skills"
-        )
-
-    def test_custom_graph_has_local_version(self, tmp_path, monkeypatch):
-        """custom=True graph should have version 'local-custom'."""
-        root = _make_registry(tmp_path)
-
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "my-skill",
-                     name="My Skill")
-        monkeypatch.chdir(tmp_path)
-
-        out_path, _ = graph_mod.write_graph_artifact(
-            root, fmt="json", custom=True
-        )
-
-        data = json.loads(out_path.read_text(encoding="utf-8"))
-        assert data["version"] == "local-custom", (
-            "Custom graph version should be 'local-custom'"
-        )
-
-
-# ═══════════════════════════════════════════════════════════════════════════
-# GREEN TESTS — scenarios that PASS on the PR branch
-# ═══════════════════════════════════════════════════════════════════════════
-
-
-class TestGreen_ScanGlobalSearch:
-    """GREEN #1-2: scan_skill_mds global_search parameter works correctly."""
-
-    def test_local_scan_finds_project_skills(self, tmp_path):
-        """global_search=False finds skills under the project root."""
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "project-skill",
-                     name="Project Skill")
-
-        results = scan_skill_mds(root=str(tmp_path), global_search=False)
-        found_ids = {sk["id"] for sk in results}
-        # scan_skill_mds returns slash-prefixed IDs
-        assert "/project-skill" in found_ids
-
-    def test_global_search_includes_global_dirs(self, tmp_path, monkeypatch):
-        """global_search=True includes global user skill directories."""
-        home = tmp_path / "fake-home"
-        global_skill = home / ".agents" / "skills" / "global-skill"
-        global_skill.mkdir(parents=True)
-        _write(str(global_skill / "skill.md"),
-               "---\nname: Global Skill\ndescription: Installed globally\n---\n")
-
-        monkeypatch.setenv("HOME", str(home))
-
-        dirs = _skill_search_dirs(root=str(tmp_path), global_search=True)
-        real_dirs = [os.path.realpath(d) for d in dirs]
-        global_path = os.path.realpath(str(home / ".agents" / "skills"))
-        assert global_path in real_dirs
-
-
-class TestGreen_PrerequisitesAlwaysPresent:
-    """GREEN #3: scan_skill_mds result always has 'prerequisites' key."""
-
-    def test_prerequisites_populated_from_frontmatter(self, tmp_path):
-        """Skills with frontmatter prerequisites have them in the result."""
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "advanced",
-                     name="Advanced", prerequisites=["basic-skill"])
-
-        results = scan_skill_mds(root=str(tmp_path))
-        assert len(results) == 1
-        # Prerequisites parsed from frontmatter are returned as a string
-        # (since _read_skill_md does simple line parsing, not full YAML)
-        prereqs = results[0]["prerequisites"]
-        assert prereqs is not None, "prerequisites key should be present"
-
-    def test_prerequisites_empty_when_absent(self, tmp_path):
-        """Skills without prerequisites frontmatter get empty list."""
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "basic",
-                     name="Basic Skill")
-
-        results = scan_skill_mds(root=str(tmp_path))
-        assert len(results) == 1
-        assert results[0]["prerequisites"] == []
-
-
-class TestGreen_PushInjection:
-    """GREEN #4: push_command injection logic works correctly."""
-
-    def test_injected_skill_has_correct_fields(self, tmp_path):
-        """Verify the shape of injected proposed skills."""
-        batch = {"sourceRepo": "test/repo", "proposedSkills": [], "knownSkills": []}
-        batch_proposed_ids = set()
-        batch_known_ids = set()
-
-        installed_skills = [
-            {"id": "new-skill", "name": "New Skill", "description": "Something new"},
-        ]
-
-        for sk in installed_skills:
-            cid = sk["id"]
-            if cid not in batch_proposed_ids and cid not in batch_known_ids:
-                batch.setdefault("proposedSkills", []).append({
-                    "id": cid,
-                    "name": sk.get("name", cid),
-                    "type": "basic",
-                    "description": sk.get("description", f"Local custom skill {cid}"),
-                    "sourceRepo": batch.get("sourceRepo", "unknown"),
-                    "lifecycle": "pending",
-                })
-                batch_proposed_ids.add(cid)
-
-        injected = batch["proposedSkills"][0]
-        assert injected["id"] == "new-skill"
-        assert injected["name"] == "New Skill"
-        assert injected["type"] == "basic"
-        assert injected["description"] == "Something new"
-        assert injected["sourceRepo"] == "test/repo"
-        assert injected["lifecycle"] == "pending"
-
-
-class TestGreen_CustomGraphMatchesScan:
-    """GREEN #5: write_graph_artifact(custom=True) graph 'skills' list matches scan output."""
-
-    def test_custom_graph_nodes_match_scan(self, tmp_path, monkeypatch):
-        """Custom graph nodes should exactly match scan_skill_mds output IDs."""
-        root = _make_registry(tmp_path)
-
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "alpha",
-                     name="Alpha", description="First")
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "beta",
-                     name="Beta", description="Second")
-        monkeypatch.chdir(tmp_path)
-
-        # Run scan to get expected IDs (slash-prefixed from scan_skill_mds)
-        scanned = scan_skill_mds(root=str(tmp_path), global_search=False)
-        expected_ids = {sk["id"].lstrip("/") for sk in scanned}
-
-        # Run custom graph
-        out_path, _ = graph_mod.write_graph_artifact(
-            root, fmt="json", custom=True
-        )
-        data = json.loads(out_path.read_text(encoding="utf-8"))
-        node_ids = {n["id"] for n in data["nodes"]}
-
-        assert node_ids == expected_ids, (
-            f"Custom graph nodes {node_ids} should match scan output {expected_ids}"
-        )
-
-    def test_custom_graph_preserves_prerequisites_as_edges(self, tmp_path, monkeypatch):
-        """Prerequisites from scan_skill_mds should become edges in the custom graph."""
-        root = _make_registry(tmp_path)
-
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "parent-skill",
-                     name="Parent")
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "child-skill",
-                     name="Child", prerequisites=["parent-skill"])
-        monkeypatch.chdir(tmp_path)
-
-        out_path, _ = graph_mod.write_graph_artifact(
-            root, fmt="json", custom=True
-        )
-        data = json.loads(out_path.read_text(encoding="utf-8"))
-
-        # Check that an edge exists from parent-skill to child-skill
-        # Note: prerequisites in frontmatter are parsed as a string by _read_skill_md,
-        # not as a list. The custom graph code calls fm.get("prerequisites", [])
-        # which may return the raw string. Let's verify actual edge behavior.
-        edges = data.get("edges", [])
-        # Edges depend on whether prerequisites parse correctly as a list
-        # from the simple frontmatter parser. This validates the integration.
-        node_ids = {n["id"] for n in data["nodes"]}
-        assert "parent-skill" in node_ids
-        assert "child-skill" in node_ids
-
-
-# ═══════════════════════════════════════════════════════════════════════════
-# SCRUTINY TESTS — Key areas flagged for deeper inspection
-# ═══════════════════════════════════════════════════════════════════════════
-
-
-class TestScrutiny_PushDuplicateGuard:
-    """Scrutiny #1: pushable_custom_skills = installed_skills.
-
-    Verify that the interactive exclusion loop correctly removes skills
-    and that no duplicates are introduced if a skill was already in
-    batch_proposed_ids before the injection loop runs.
-    """
-
-    def test_no_duplicates_after_injection_and_exclusion(self, tmp_path):
-        """If a skill is already proposed, injection should not duplicate it."""
-        batch = {
-            "sourceRepo": "test/repo",
-            "proposedSkills": [
-                {"id": "existing", "name": "Existing", "type": "basic",
-                 "description": "", "sourceRepo": "test/repo", "lifecycle": "pending"},
-            ],
-            "knownSkills": [],
-        }
-        batch_proposed_ids = {s["id"] for s in batch["proposedSkills"]}
-        batch_known_ids = set()
-
-        installed_skills = [
-            {"id": "existing", "name": "Existing", "description": ""},
-            {"id": "new-one", "name": "New One", "description": ""},
-        ]
-
-        # Run the injection logic
-        for sk in installed_skills:
-            cid = sk["id"]
-            if cid not in batch_proposed_ids and cid not in batch_known_ids:
-                batch.setdefault("proposedSkills", []).append({
-                    "id": cid, "name": sk.get("name", cid), "type": "basic",
-                    "description": sk.get("description", ""),
-                    "sourceRepo": batch.get("sourceRepo", "unknown"),
-                    "lifecycle": "pending",
-                })
-                batch_proposed_ids.add(cid)
-
-        # Count occurrences
-        existing_count = sum(1 for s in batch["proposedSkills"] if s["id"] == "existing")
-        assert existing_count == 1, (
-            f"'existing' appears {existing_count} times in proposedSkills — expected 1"
-        )
-
-        # Simulate exclusion
-        excluded_ids = {"new-one"}
-        batch["proposedSkills"] = [
-            s for s in batch["proposedSkills"] if s["id"] not in excluded_ids
-        ]
-        remaining_ids = {s["id"] for s in batch["proposedSkills"]}
-        assert "new-one" not in remaining_ids
-        assert "existing" in remaining_ids
-
-
-class TestScrutiny_ShowTreeCustomMode:
-    """Scrutiny #2: show_tree custom mode calls scan_skill_mds without root.
-
-    Verify show_tree(custom=True) correctly filters to local custom skills.
-    """
-
-    def test_custom_mode_shows_only_custom_skills(self, tmp_path, monkeypatch, capsys):
-        """custom=True should filter display to local custom / non-canonical skills."""
-        monkeypatch.chdir(tmp_path)
-
-        # Create a local custom skill
-        _make_skill(tmp_path, os.path.join(".agents", "skills"), "my-custom",
-                     name="My Custom Skill")
-
-        graph_data = {
-            "skills": [
-                {"id": "web-search", "name": "Web Search", "type": "basic",
-                 "level": "1★", "prerequisites": []},
-            ]
-        }
-
-        tree_data = {
-            "userId": "testuser",
-            "updatedAt": "2026-01-01",
-            "unlockedSkills": [
-                {"skillId": "web-search", "level": "1★"},
-                {"skillId": "my-custom", "level": "0★"},
-            ],
-            "pendingCombinations": [],
-            "stats": {},
-        }
-
-        show_tree(tree_data, graph_data=graph_data,
-                  registry_path=str(tmp_path), custom=True)
-        out = capsys.readouterr().out
-
-        # my-custom should appear (it's a local custom skill)
-        assert "my-custom" in out, "Custom skill should be shown in custom mode"
-
-        # web-search is canonical and NOT in local_custom_ids — it should be hidden
-        assert "web-search" not in out, (
-            "Canonical-only skill should not appear in custom mode"
-        )
-
-
-class TestScrutiny_CustomGraphSchema:
-    """Scrutiny #3: custom graph schema consumed by build_render_graph.
-
-    The synthetic graph dict uses 'version': 'local-custom' and a flat
-    'skills' list. Verify build_render_graph can consume this without errors.
-    """
-
-    def test_build_render_graph_handles_custom_schema(self):
-        """build_render_graph should not crash on the custom graph schema."""
-        custom_graph = {
-            "version": "local-custom",
-            "skills": [
-                {"id": "my-skill", "name": "My Skill", "type": "basic",
-                 "level": "0★", "prerequisites": []},
-                {"id": "my-other", "name": "Other Skill", "type": "basic",
-                 "level": "0★", "prerequisites": ["my-skill"]},
-            ],
-        }
-
-        render_graph = graph_mod.build_render_graph(custom_graph)
-
-        # Should produce valid nodes
-        assert len(render_graph["nodes"]) == 2
-        node_ids = {n["id"] for n in render_graph["nodes"]}
-        assert node_ids == {"my-skill", "my-other"}
-
-        # Should produce the prerequisite edge
-        assert len(render_graph["edges"]) == 1
-        assert render_graph["edges"][0]["source"] == "my-skill"
-        assert render_graph["edges"][0]["target"] == "my-other"
-
-        # Version should carry through
-        assert render_graph["version"] == "local-custom"
-
-
-class TestScrutiny_ScanRootVsParent:
-    """Scrutiny #4: scan root change — root vs root/..
-
-    The old behaviour walked the parent of root. The new behaviour (when
-    global_search=False) walks root directly. Verify edge case: skills one
-    level above the project root are not found.
-    """
-
-    def test_skills_above_project_root_not_found(self, tmp_path):
-        """Skills in the parent of root should NOT be found with default scan."""
-        parent_skills = tmp_path / ".agents" / "skills" / "parent-skill"
-        parent_skills.mkdir(parents=True)
-        _write(str(parent_skills / "skill.md"),
-               "---\nname: Parent Skill\ndescription: lives above project\n---\n")
-
-        project = tmp_path / "my-project"
-        project.mkdir()
-
-        results = scan_skill_mds(root=str(project), global_search=False)
-        found_ids = {r["id"] for r in results}
-        assert "parent-skill" not in found_ids, (
-            "Skills above the project root should not be found with global_search=False"
-        )
-
-
-class TestScrutiny_AllFlagShadowing:
-    """Scrutiny #5: --all flag naming.
-
-    args.all shadows the Python builtin `all`. Verify the parser attribute
-    is properly accessible and doesn't cause runtime issues.
-    """
-
-    def test_all_flag_accessible_on_namespace(self):
-        """Verify that an argparse Namespace with 'all' attribute works correctly."""
-        ns = SimpleNamespace(all=True)
-        assert getattr(ns, "all", False) is True
-
-        ns2 = SimpleNamespace(all=False)
-        assert getattr(ns2, "all", False) is False
-
-    def test_parser_produces_all_attribute(self):
-        """Verify the actual parser produces args.all for the scan command."""
-        from gaia_cli.main import get_parser
-        parser, _ = get_parser()
-        args = parser.parse_args(["scan", "--all"])
-        assert getattr(args, "all", False) is True
-
-    def test_parser_custom_on_tree(self):
-        """Verify --custom flag on tree command."""
-        from gaia_cli.main import get_parser
-        parser, _ = get_parser()
-        args = parser.parse_args(["tree", "--custom"])
-        assert getattr(args, "custom", False) is True
-
-    def test_parser_custom_on_graph(self):
-        """Verify --custom flag on graph command."""
-        from gaia_cli.main import get_parser
-        parser, _ = get_parser()
-        args = parser.parse_args(["graph", "--custom"])
-        assert getattr(args, "custom", False) is True
-
-    def test_parser_canon_on_tree(self):
-        """Verify --canon flag on tree command."""
-        from gaia_cli.main import get_parser
-        parser, _ = get_parser()
-        args = parser.parse_args(["tree", "--canon"])
-        assert getattr(args, "canon", False) is True
-
-    def test_parser_canon_on_graph(self):
-        """Verify --canon flag on graph command."""
-        from gaia_cli.main import get_parser
-        parser, _ = get_parser()
-        args = parser.parse_args(["graph", "--canon"])
-        assert getattr(args, "canon", False) is True
