+++ tests/test_push.py
@@ -252,5 +252,158 @@ class TestGaiaPush(unittest.TestCase):
                 )
 
 
+# ═══════════════════════════════════════════════════════════════════════════
+# Relocated from test_pr635_review.py — push injection logic
+# ═══════════════════════════════════════════════════════════════════════════
+
+
+class TestRed_PushInjectsUnmatchedSkills:
+    """RED #4: push_command injects unmatched local custom skills into batch.
+
+    On main: skills not already in batch_proposed_ids or batch_known_ids were DROPPED.
+    On branch: they are injected into batch['proposedSkills'].
+
+    This test verifies the injection logic directly rather than running the
+    full push_command (which requires interactive input and git remotes).
+    """
+
+    def test_injection_logic_adds_missing_skills_to_batch(self, tmp_path):
+        """Simulate the push_command injection loop: skills not in batch should be added."""
+        # Simulate what push_command does
+        batch = {
+            "sourceRepo": "test/repo",
+            "proposedSkills": [
+                {"id": "already-proposed", "name": "Already", "type": "basic",
+                 "description": "Already in batch", "sourceRepo": "test/repo",
+                 "lifecycle": "pending"}
+            ],
+            "knownSkills": [
+                {"skillId": "already-known"}
+            ],
+        }
+        batch_proposed_ids = {s["id"] for s in batch.get("proposedSkills", [])}
+        batch_known_ids = {s["skillId"] for s in batch.get("knownSkills", [])}
+
+        # Simulate installed_skills (what scan_skill_mds returns)
+        installed_skills = [
+            {"id": "already-proposed", "name": "Already Proposed", "description": ""},
+            {"id": "already-known", "name": "Already Known", "description": ""},
+            {"id": "brand-new-skill", "name": "Brand New", "description": "A new skill"},
+        ]
+
+        # This is the injection logic from the PR branch
+        for sk in installed_skills:
+            cid = sk["id"]
+            if cid not in batch_proposed_ids and cid not in batch_known_ids:
+                batch.setdefault("proposedSkills", []).append({
+                    "id": cid,
+                    "name": sk.get("name", cid),
+                    "type": "basic",
+                    "description": sk.get("description", f"Local custom skill {cid}"),
+                    "sourceRepo": batch.get("sourceRepo", "unknown"),
+                    "lifecycle": "pending",
+                })
+                batch_proposed_ids.add(cid)
+
+        # Verify: brand-new-skill should have been injected
+        proposed_ids = {s["id"] for s in batch["proposedSkills"]}
+        assert "brand-new-skill" in proposed_ids, (
+            "Unmatched local custom skill should be injected into proposedSkills"
+        )
+        # Verify no duplicates
+        assert len([s for s in batch["proposedSkills"] if s["id"] == "already-proposed"]) == 1, (
+            "Already-proposed skill should not be duplicated"
+        )
+
+
+class TestGreen_PushInjection:
+    """GREEN #4: push_command injection logic works correctly."""
+
+    def test_injected_skill_has_correct_fields(self, tmp_path):
+        """Verify the shape of injected proposed skills."""
+        batch = {"sourceRepo": "test/repo", "proposedSkills": [], "knownSkills": []}
+        batch_proposed_ids = set()
+        batch_known_ids = set()
+
+        installed_skills = [
+            {"id": "new-skill", "name": "New Skill", "description": "Something new"},
+        ]
+
+        for sk in installed_skills:
+            cid = sk["id"]
+            if cid not in batch_proposed_ids and cid not in batch_known_ids:
+                batch.setdefault("proposedSkills", []).append({
+                    "id": cid,
+                    "name": sk.get("name", cid),
+                    "type": "basic",
+                    "description": sk.get("description", f"Local custom skill {cid}"),
+                    "sourceRepo": batch.get("sourceRepo", "unknown"),
+                    "lifecycle": "pending",
+                })
+                batch_proposed_ids.add(cid)
+
+        injected = batch["proposedSkills"][0]
+        assert injected["id"] == "new-skill"
+        assert injected["name"] == "New Skill"
+        assert injected["type"] == "basic"
+        assert injected["description"] == "Something new"
+        assert injected["sourceRepo"] == "test/repo"
+        assert injected["lifecycle"] == "pending"
+
+
+class TestScrutiny_PushDuplicateGuard:
+    """Scrutiny #1: pushable_custom_skills = installed_skills.
+
+    Verify that the interactive exclusion loop correctly removes skills
+    and that no duplicates are introduced if a skill was already in
+    batch_proposed_ids before the injection loop runs.
+    """
+
+    def test_no_duplicates_after_injection_and_exclusion(self, tmp_path):
+        """If a skill is already proposed, injection should not duplicate it."""
+        batch = {
+            "sourceRepo": "test/repo",
+            "proposedSkills": [
+                {"id": "existing", "name": "Existing", "type": "basic",
+                 "description": "", "sourceRepo": "test/repo", "lifecycle": "pending"},
+            ],
+            "knownSkills": [],
+        }
+        batch_proposed_ids = {s["id"] for s in batch["proposedSkills"]}
+        batch_known_ids = set()
+
+        installed_skills = [
+            {"id": "existing", "name": "Existing", "description": ""},
+            {"id": "new-one", "name": "New One", "description": ""},
+        ]
+
+        # Run the injection logic
+        for sk in installed_skills:
+            cid = sk["id"]
+            if cid not in batch_proposed_ids and cid not in batch_known_ids:
+                batch.setdefault("proposedSkills", []).append({
+                    "id": cid, "name": sk.get("name", cid), "type": "basic",
+                    "description": sk.get("description", ""),
+                    "sourceRepo": batch.get("sourceRepo", "unknown"),
+                    "lifecycle": "pending",
+                })
+                batch_proposed_ids.add(cid)
+
+        # Count occurrences
+        existing_count = sum(1 for s in batch["proposedSkills"] if s["id"] == "existing")
+        assert existing_count == 1, (
+            f"'existing' appears {existing_count} times in proposedSkills — expected 1"
+        )
+
+        # Simulate exclusion
+        excluded_ids = {"new-one"}
+        batch["proposedSkills"] = [
+            s for s in batch["proposedSkills"] if s["id"] not in excluded_ids
+        ]
+        remaining_ids = {s["id"] for s in batch["proposedSkills"]}
+        assert "new-one" not in remaining_ids
+        assert "existing" in remaining_ids
+
+
 if __name__ == "__main__":
     unittest.main()
